fix: moving constants from api and init file to a separate constants file
