+++ lms/djangoapps/instructor/constants.py
@@ -1,0 +1,9 @@
+"""
+Constants used by Instructor.
+"""
+
+# this is the UserPreference key for the user's recipient invoice copy
+INVOICE_KEY = 'pref-invoice-copy'
+
+# external plugins (if any) will use this constant to return context to instructor dashboard
+INSTRUCTOR_DASHBOARD_PLUGIN_VIEW_NAME = 'instructor_dashboard'
