--- lms/djangoapps/instructor/views/__init__.py
@@ -1,4 +1,0 @@
-"""
-This is the UserPreference key for the user's recipient invoice copy
-"""
-INVOICE_KEY = 'pref-invoice-copy'
